Added color changing grid item
Added the ability for the user to click on a grid item and have it change to the _plantColor before linearly changing back to its original color using the Color.Lerp function. The current issue is that it functions based off of a bool which I do not yet know how to change back. In short, the grid item can change color but only ONCE.